feat: leave history impl

+++ lib/presentation/leaves_and_rebate/bloc/leaves_and_rebate_bloc.dart
@@ -26,11 +26,11 @@ class LeavesAndRebateBloc
   FutureOr<void> _onFetchLeavesAndRebates(
       FetchLeavesAndRebates event, Emitter<LeavesAndRebateState> emit) async {
     PaginatedLeaves currYearLeaves =
-        await leaveRepository.getLeaves(DateTime.now().year, 0);
+        await leaveRepository.getLeaves(DateTime.now().year.toInt(), 0);
     int remainingLeaves = await leaveRepository.remainingLeaves();
     PaginatedYearlyRebate initialYearlyRebates =
         await transactionRepository.getYearlyRebates(
-            DateTime(DateTime.now().year, DateTime.now().month - 1).year);
+            DateTime(DateTime.now().year, DateTime.now().month - 1).year);       
     emit(LeavesAndRebateState(
       remainingLeaves: remainingLeaves,
       mealsSkipped: maxLeaves - remainingLeaves,
